Split tmux session control client tests
